Simple changes for Qt4 compatibility
These are changes that are compatible with Qt5, so don't need
defines.  The question querybox has changed the default interface,
so be explicit; a combobox method is new, but does the same (in
this case) as an existing method, so use that; the QUrl method is
new but is doing the same as the old ::path method in this case
(because raw_name here is just the last path component).

Signed-off-by: Michael Matz <matz@suse.de>