Add conditional compatibility with Qt4
This adds some compatibility code that needs QT_VERSION checks.
The tree model doesn't yet have Qt::ItemNeverHasChildren, or
fine grained roles for changes, and no sections.  The message
handler interface is different (and doesn't provide a context).
In the selection model you can't clear the current index alone,
but that can be done with setting it to a default one (which is
actually the implementation in Qt5 as well).

Signed-off-by: Michael Matz <matz@suse.de>