[qt4] Adjust empty TARGET in *.pro
The Qt4 qmake doesn't like an really empty TARGET specification
in *.pro.  It replaces it by the name of the containing subdir,
hence a make in e.g. 'scripts/' tries to make the executable
'scripts', of course without defined SOURCES that fails.

Trick qmake by making it look like TARGET is not empty, when
in reality it will be by using a not-defined make variable.

Signed-off-by: Michael Matz <matz@suse.de>