Add default applications for macOS
I leave the desktop detection code unchanged because in theory it's
possible to run X11 with KDE/GNOME/whatever on Darwin, so just fall back
to the normal mac desktop and use the system applications.

The terminal is a bit complicated because there's no easy way to read
out the default terminal with one line of code. Qt creator uses a
AppleScript for this. However, opening the default macOS terminal is
still better than nothing, and if the user really wants to use iTerm
(for example), he can still overwrite the settings manually.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>