write in-progress output to a .part file
Ensure that the qdirstat cache file is the previous output,
or the refreshed output, but never partial/corrupt from
a current operation in progress or from an aborted run.

The cache writing process can take a considerable amount of
time, giving a large window when the server could be rebooted,
power outage, etc. leaving the user with neither the old
cache file nor the new one.