Remember the two special categories
Keep pointers to the executable category and symlink category to avoid searching for them every time and save a few cycles.  They are fairly common cases and this code already makes up a good proportion of the initial treemap paint, although mostly in the string handling a regular expression matches.