Mime categorizer config fixes
Override the ListEditor updateActions() so that the remove button remains disabled for the symlink and executable categories when they become the current item after a category has been deleted.

Disable everything except the add button when there are no categories (shouldn't happen but was happening in some cases before this fix).  Also set the mini-treemap to the default fixed colour when the last category is removed, and reset the name, colour and patterns in the dialog.