Fix timer not stopped on abort
The DirReadQueue zero-second timer is not stopped when a read is aborted, which leads to 100% cpu usage and lockup of some gui actions.  The timer is only stopped when a job completes normally and at that point the queue is empty.

The timer could just be stopped in clear(), but to catch any other cases now or in the future, stop the timer whenever the queue is empty at the time that the timer fires.